Replace the back link panel with a jinja extension
