+++ h/templates/groups/edit.html.jinja2
@@ -7,6 +7,6 @@
   {{ form }}
 
   <footer class="form-footer">
-    {{ panel('back_link') }}
+    {% include "h:templates/includes/back_link.html.jinja2" %}
   </footer>
 {% endblock page_content %}
