+++ h/templates/layouts/account.html.jinja2
@@ -32,7 +32,7 @@
       {{ self.page_content() }}
 
       <footer class="form-footer--no-border">
-        {{ panel('back_link') }}
+        {% include "h:templates/includes/back_link.html.jinja2" %}
       </footer>
     </div>
   </div>
