--- h/templates/panels/back_link.html.jinja2
@@ -1,4 +1,0 @@
-{% if back_location and back_label %}
-<a class="link--footer"
-   href="{{ back_location }}">← {{ back_label }}</a>
-{% endif %}
